eyevory_backend/requirements.txt
asgiref==3.5.0
autopep8==1.6.0
certifi==2021.10.8
Django==4.0.4
django-cors-headers==3.11.0
django-filter==21.1
djangorestframework==3.13.1
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.1.0
influxdb-client==1.28.0
Markdown==3.3.6
pycodestyle==2.8.0
PyJWT==1.7.1
python-dateutil==2.8.2
pytz==2022.1
Rx==3.2.0
six==1.16.0
sqlparse==0.4.2
toml==0.10.2
urllib3==1.26.9
